Material Development Library
Going to try a new css library. Going to see if it looks really nice and will compare on mobile as well. Going to try and get something that I can really use as a web portal for not just myself, but also others. Might be a new homepage coming for spollock.xyz as well if I can swing it